BS-LDM/requirements.txt
numpy~=1.26.0
torchvision~=0.15.2
opencv-python~=4.8.1.78
pandas~=2.1.3
matplotlib~=3.8.0
pillow~=10.0.1
diffusers~=0.23.1
tqdm~=4.65.0
monai~=1.3.0
scikit-image~=0.22.0
lpips~=0.1.4
monai-generative~=0.2.2
torch_ema~=0.3